+++ src/main/java/dtalk/dto/user/UserFindDTO.java
@@ -14,7 +14,7 @@ public class UserFindDTO {
     private String nickname;
     private String profileImg;
     private Integer quizCount;
-    public static UserFindDTO createUserFindDto(User user,Integer quizCount){
+    public static UserFindDTO createUserFindDto(User user,Long quizCount){
         UserFindDTO userFindDto = new UserFindDTO(user.getIdx(), user.getId(), user.getNickname(), user.getProfileImg(),quizCount);
         return userFindDto;
     }
